Fixed an ObjectDisposedException when generating a HiLo id.
The static CollectionHiLoIdGenerator was caching HiLoGenerator objects for
collections which in turn cached DB connections that could have been
closed. Closing and reopening a connection and inserting in the same
collection sometimes crashed because of that.

I've kept the static, but it no longer keeps a connection around.I changed
the GenerateId method to accept a connection as a parameter.